[sofa-launcher] Add basic pythons scripts & tools to start multiple runSofa in parallel.
The tools takes a set of sofa scene and execute them with runSofa in batch mode.

The execution can be done sequentially or in parallel on a local machine or on a
cluster of machines if distributed.

Additionnally the scenes and script controller can be automatically generated from
kind of templated scenes  using the Cheethat template language.

With this tool it become very easy to simulate a single scene varying some
parameters and gather the results for comparison purpose.

This is a Work in Progress...